% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_criteria.R
\name{check_stop_criteria}
\alias{check_stop_criteria}
\title{Stop criteria for DE}
\usage{
check_stop_criteria(L)
}
\value{
logical flag indicating whether any stop condition has been reached.
}
\description{
Implements different stop criteria for the ExpDE framework
}
