% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate_population.R
\name{evaluate_population}
\alias{evaluate_population}
\title{Evaluate DE population}
\usage{
evaluate_population(probpars, Pop)
}
\arguments{
\item{probpars}{problem parameters (see \code{\link{ExpDE}} for details).}

\item{Pop}{population matrix (each row is a candidate solution, normalized
to the [0, 1] interval,)}
}
\value{
numeric vector (with length \code{nrow(Pop)}) containing the function
values of each point in the population.
}
\description{
Evaluates the DE population on a given objective function.
}

