% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpDE.R
\name{ExpDE}
\alias{ExpDE}
\title{Differential Evolution Algorithm (DE)}
\usage{
ExpDE(popsize = 30, mutName = c("rand1", "rand2"), mutF = 0.2,
  recName = c("bin", "exp"), selName = c("standard", "other"),
  convType = c("niter", "stab"), convNiter = 500, convStab = 5,
  probName = c("myfun", "other"), lim_inf = -5.12, lim_sup = 5.12,
  probOpt = 3)
}
\arguments{
\item{popsize}{size of population.}

\item{mutName}{is name of mutation method name;}

\item{mutF}{is the scaling factor.}

\item{recName}{is the type of recombination to be used in the algorithm.}

\item{selName}{is the type of selection to be in the algorithm.}

\item{convType}{is name of convergence criteria for the algorithm.}

\item{convStab}{is value of stabilization algorithm.}

\item{lim_inf}{is lower limit of the information to be generated.}

\item{lim_sup}{is upper limit of the information to be generated.}

\item{probOpt}{is other parameters to the algorithm.}

\item{conNiter}{is value of convergence criteria for the algorithm.}
}
\value{
matrix with result of the optimization.
}
\description{
Is a method that optimizes a problem by iteratively trying to
improve a candidate solution with regard to a given measure of quality
}
\examples{
ExpDE(popsize = 50, mutName ="rand", mutF = 0.3), recName = "bin",
 selName = "standard", convType = "niter", convNiter = 400, probName = "myfun",
 lim_inf = -5.12, lim_sup = 5.12, probOpt = 3)
}
\keyword{de,}
\keyword{optimization}

