% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_methods.R
\name{gen_methods}
\alias{gen_methods}
\title{Generate methods for testing}
\usage{
gen_methods(ID, echo = TRUE)
}
\arguments{
\item{ID}{a unique positive integer (e.g., a student ID)}

\item{echo}{Boolean, should the function print its output?}
}
\value{
List object with a number of configurations to use with \code{\link{ExpDE}}. Alternatively, 
the method names returned by this function can be passed directly to \code{\link{ExpDE2}}.
}
\description{
Generate a number of random DE configurations.
}
