% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_problems.R
\name{gen_problems}
\alias{gen_problems}
\title{Generate test problems}
\usage{
gen_problems(ID, nprobs, echo = TRUE)
}
\arguments{
\item{ID}{a unique positive integer (e.g., a student ID)}

\item{nprobs}{number of problems to generate}

\item{echo}{Boolean, should the function print its output?}
}
\value{
List object with definitions for nprobs test problems.
}
\description{
Get a set of test problems for practicing the design and running of 
comparative experiments using ExpDE as a data-generating process.
}
