% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get_exp_params_1.R
\name{get_exp_params_1}
\alias{get_exp_params_1}
\title{Get experimental parameters}
\usage{
get_exp_params_1(ID)
}
\arguments{
\item{ID}{a unique positive integer (e.g., a student ID)}
}
\value{
List object with experimental parameters that need to be used in the 
design and execution of the experiment.
}
\description{
Get parameters for practicing the design and running of a factorial experiment 
using ExpDE as a data-generating process.
}
