% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get_experimental_parameters.R
\name{get_experimental_parameters}
\alias{get_experimental_parameters}
\title{Get experimental parameters}
\usage{
get_experimental_parameters(ID)
}
\arguments{
\item{ID}{student ID}
}
\value{
List object with experimental parameters that need to be used in the 
design and execution of the experiment.
}
\description{
Get parameters for practicing the design and running of a factorial experiment 
using ExpDE as a data-generating process.
}
