% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_jade.R
\name{mutation_jade}
\alias{mutation_jade}
\title{/current-to-pbest mutation for adaptive DE}
\usage{
mutation_jade(L, mutpars)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}

\item{mutpars}{mutation parameters (see \code{Mutation parameters} for details)}
}
\value{
Matrix \code{M} containing the mutated population
}
\description{
Implements the "/current-to-pbest" mutation with changes to the use
of the JADE self-adaptation method
}
\details{
This routine also implements one special case: 
\itemize{
 \item current-to-best mutation (\code{current_to_best}), by setting 
   \code{mutpars$p = 1}); 
 \item Flat recombination (\code{flat}), by setting 
   \code{recpars$alpha = recpars$beta = 0})
}
}
\section{Mutation Parameters}{

The \code{mutpars} parameter contains all parameters required to define the 
mutation. \code{mutation_jade()} understands the following fields in 
\code{mutpars}:
\itemize{
   \item \code{p} : either the number of "best" vectors to use (if given as a 
   positive integer) or the proportion of the population to use as "best"
   vectors (if 0 < p < 1). \cr
}
\code{adapars}:
\itemize{
   \item \code{Fi} : set of scaling factor of each individual for difference
    vector(s). 
}
}

\section{Warning}{

This routine will search for the performance vector 
of population \code{X} (\code{J}) in the list \code{L}. This 
variable must be defined for \code{mutation_jade()} to work.
}

\section{X}{

Population matrix (original).
}

\section{J}{

Performance vector for population \code{X}.
}

\section{References}{

J. Zhang, A.C. Sanderson, 
"JADE: Adaptive differential evolution with optional external archive". 
IEEE Transactions on Evolutionary Computation 13:945-958, 2009
}

\author{
Felipe Campelo (\email{fcampelo@ufmg.br})
}
