% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_mean.R
\name{mutation_mean}
\alias{mutation_mean}
\title{/mean mutation for DE}
\usage{
mutation_mean(L, mutpars)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}

\item{mutpars}{mutation parameters (see \code{Mutation parameters} for details)}
}
\value{
Matrix \code{M} containing the mutated population
}
\description{
Implements the "/mean/nvecs" mutation for the ExpDE framework
}
\section{Mutation Parameters}{

The \code{mutpars} parameter contains all parameters required to define the 
mutation. \code{mutation_mean()} understands the following fields in 
\code{mutpars}:
\itemize{
   \item \code{f} : scaling factor for difference vector(s).\cr
   Accepts numeric vectors of size 1 or \code{nvecs}.
   \item \code{nvecs} : number of difference vectors to use.\cr 
       Accepts \code{1 <= nvecs <= (nrow(X)/2 - 2)}\cr
       Defaults to 1.
}
}

\section{X}{

Population matrix (original).
}

\section{References}{

K. Price, R.M. Storn, J.A. Lampinen, "Differential Evolution: A 
Practical Approach to Global Optimization", Springer 2005
}

\author{
Felipe Campelo (\email{fcampelo@ufmg.br})
}
