% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_none.R
\name{mutation_none}
\alias{mutation_none}
\title{NULL mutation for DE}
\usage{
mutation_none(L, mutpars)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}

\item{mutpars}{mutation parameters (see \code{Mutation parameters} for details)}
}
\value{
@return The same matrix \code{X} used as an input.
}
\description{
Implements the "/none" mutation (i.e., no mutation performed) for the 
ExpDE framework
}
\section{Mutation Parameters}{

The \code{mutpars} parameter contains all parameters required to define the 
mutation. \code{mutation_none()} requires no fields in this parameter.
}

