% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_operators.R
\name{mutation_operators}
\alias{mutation_operators}
\title{Mutation operators available}
\usage{
mutation_operators()
}
\value{
Character vector with the names of all mutation operators
}
\description{
List all available mutation operators in the ExpDE package
}

