% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_mutation.R
\name{perform_mutation}
\alias{perform_mutation}
\title{Perform mutation}
\usage{
perform_mutation(L)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}
}
\value{
List \code{L} containing all updated parameters, including
the mutated population \code{M}
}
\description{
Implements the control of mutation functions for the ExpDE framework
}
\section{Mutation Parameters}{

The \code{mutpars} parameter contains all parameters required to define the 
mutation.
}

\author{
Bruna Queiroz (\email{bsqueiroz98@gmail.com})
}
