% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_recombination.R
\name{perform_recombination}
\alias{perform_recombination}
\title{Perform recombination}
\usage{
perform_recombination(L)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}
}
\value{
List \code{L} containing all updated parameters, including
the matrix \code{U} containing the recombined population
}
\description{
Implements the control of recombination functions for the ExpDE framework
}
\section{Recombination Parameters}{

The \code{recpars} parameter contains all parameters required to define the 
recombination.
}

\author{
Bruna Queiroz (\email{bsqueiroz98@gmail.com})
}
