% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_selection.R
\name{perform_selection}
\alias{perform_selection}
\title{Perform selection}
\usage{
perform_selection(L)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}
}
\value{
List \code{L} containing all updated parameters, including
the mutated population \code{M}
}
\description{
Implements the control of selection functions for the ExpDE framework
}
\section{Selection Parameters}{

The \code{selpars} parameter contains all parameters required to define the 
selection.
}

\author{
Bruna Queiroz (\email{bsqueiroz98@gmail.com})
}
