% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_selfadaptive.R
\name{perform_selfadaptive}
\alias{perform_selfadaptive}
\title{Perform self-adaptive methods}
\usage{
perform_selfadaptive(L)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}
}
\value{
List \code{L} containing all updated parameters
}
\description{
Implements the control of self adaptive functions for the ExpDE framework
}
\section{Self adaptive parameters}{

The \code{adapars} parameter contains all parameters required to define the 
use and control of self adaptive variables.
}

\author{
Bruna Queiroz (\email{bsqueiroz98@gmail.com})
}
