% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_bin_jade.R
\name{recombination_bin_jade}
\alias{recombination_bin_jade}
\title{/bin recombination for adaptive DE}
\usage{
recombination_bin_jade(L, recpars)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}

\item{recpars}{recombination parameters (see \code{Recombination parameters} 
for details)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/bin" (binomial) recombination with changes to the use
of the JADE self-adaptation method
}
\section{Recombination Parameters}{

The \code{recpars} parameter contains all parameters required to define the 
recombination. \code{recombination_bin_jade()} understands the following fields in 
\code{recpars}:
\itemize{
   \item \code{minchange} : logical flag to force each new candidate solution 
                            to inherit at least one component from its 
                            mutated 'parent'.\cr
                            Defaults to TRUE
}
\code{adapars}:
\itemize{
   \item \code{CRi}: set of probability of each individual in population
                     for using the value in \code{M}. \cr
}
}

\section{References}{

K. Price, R.M. Storn, J.A. Lampinen, "Differential Evolution: A 
Practical Approach to Global Optimization", Springer 2005

J. Zhang, A.C. Sanderson, 
"JADE: Adaptive differential evolution with optional external archive". 
IEEE Transactions on Evolutionary Computation 13:945-958, 2009
}

\section{X}{

Population matrix (original).
}

\section{M}{

Population matrix (mutated).
}

