% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombination_blxAlpha.R
\name{recombination_blxAlpha}
\alias{recombination_blxAlpha}
\title{Blend Alpha recombination for DE}
\usage{
recombination_blxAlpha(X, M, recpars = list(K = NULL, cr))
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{recpars}{recombination parameters (see \code{Recombination parameters}
for details)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/blxAlpha" (Blend Alpha) recombination for the ExpDE framework
}
\section{Recombination Parameters}{

The \code{recpars} parameter contains all parameters required to define the
recombination. \code{recombination_blxAlpha()} understands the following
fields in recpars:
   - \code{cr} : component-wise probability of selection as a cut-point.
   Accepts numeric value \code{0 < cr <= 1}.
   - \code{K} : blxAlpha constant. Defaults to 0.5.
   Accepts real value \code{0 <= K <= 1}; or \code{K = NULL} for choosing 0.5 default
}

\section{References}{

F. Herrera, M. Lozano, A. M. Sanchez, "A taxonomy for the crossover
operator for real-coded genetic algorithms: an experimental study",
International Journal of Intelligent Systems 18(3) 309-338, 2003.
}

