% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombination_flat.R
\name{recombination_flat}
\alias{recombination_flat}
\title{Flat recombination for DE}
\usage{
recombination_flat(X, M, ...)
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{...}{optional parameters (unused)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/flat" (Flat) recombination for the ExpDE
framework
}
\section{References}{

Picek, S.; Jakobovic, D.; Golub, M., "On the recombination operator
in the real-coded genetic algorithms," CEC'2013, pp.3103-3110, 2013\cr
F. Herrera, M. Lozano, J.L. Verdegay, "Tackling Real-Coded Genetic
Algorithms: Operators and Tools for Behavioural Analysis",
Artificial Intelligence Review 12 265-319, 1998.
}

