% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_operators.R
\name{recombination_operators}
\alias{recombination_operators}
\title{Recombination operators available}
\usage{
recombination_operators()
}
\value{
Character vector with the names of all recombination operator 
        routines
}
\description{
List all available recombination operators in the ExpDE package
}

