% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombination_simBinary.R
\name{recombination_simBinary}
\alias{recombination_simBinary}
\title{/simBinary recombination for DE}
\usage{
recombination_simBinary(X, M, recpars)
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{recpars}{recombination parameters (see \code{Recombination parameters}
for details)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/simBinary" (Simulated Binary) recombination for the ExpDE framework
}
\section{Recombination Parameters}{

The \code{recpars} parameter contains all parameters required to define the
recombination. \code{recombination_simBinary()} understands the following field in
recpars:
   - \code{eta} : component-wise probability.
   Accepts numeric value \code{ eta > 0}.
}

\section{References}{

K. Price, R.M. Storn, J.A. Lampinen, "Differential Evolution: A
Practical Approach to Global Optimization", Springer 2005
}

