% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_jade.R
\name{selection_jade}
\alias{selection_jade}
\title{Standard selection adapted to the JADE function}
\usage{
selection_jade(L)
}
\arguments{
\item{L}{list with all parameters for ExpDE framework}
}
\value{
List \code{L} containing all updated parameters, including
list object containing the selected population (\code{Xsel}) and 
its corresponding performance values (\code{Jsel}).
}
\description{
Implements the standard selection (greedy) for the ExpDE framework
}
\section{Files}{

Implements the update of parameters \code{A, S.cr and S.F} to the 
routine JADE.
}

\section{X}{
 
Population matrix (original).
}

\section{U}{
 
Population matrix (recombined).
}

\section{J}{
 
Performance vector for population \code{X}.
}

\section{G}{
 
Performance vector for population \code{U}.
}

