% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection_standard.R
\name{selection_standard}
\alias{selection_standard}
\title{Standard selection for DE}
\usage{
selection_standard(X, U, J, G)
}
\arguments{
\item{X}{population matrix (original)}

\item{U}{population matrix (recombined)}

\item{J}{performance vector for population \code{X}}

\item{G}{performance vector for population \code{U}}
}
\value{
list object containing the selected population (\code{Xsel}) and
its corresponding performance values (\code{Jsel}).
}
\description{
Implements the standard selection (greedy) for the ExpDE framework
}
\details{
The details (if any) come here...
}

