% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population.R
\name{create_population}
\alias{create_population}
\title{Create population}
\usage{
create_population(L)
}
\arguments{
\item{L}{list with all important variables used in ExpDE.}
}
\value{
A matrix containing the population for the ExpDE
}
\description{
Creates a new population for the ExpDE framework
}
\section{Variable probpars}{

Population size.


List of named problem parameters (see \code{\link{ExpDE}}).
}

