% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpDE.R
\name{ExpDE}
\alias{ExpDE}
\title{Differential Evolution Algorithm (DE)}
\usage{
ExpDE(popsize = 40, mutpars = list(name = "rand", f = 0.2),
  recpars = list(name = "bin"), selpars = list(name = "standard"),
  convcrit = list(types = c("niter", "stab"), pars = list(niter = 500, nstab =
  5)), probpars = list(name = "myfun", lim_inf = -5.12, lim_sup = 5.12, opt =
  3))
}
\arguments{
\item{popsize}{size of population.}

\item{mutpars}{mutation the parameter list containing:
\code{name} is name of mutation; \code{f} mutation the scaling factor.}

\item{recpars}{recombination the parameter list containing:
\code{name} is the type of recombination to be used in the algorithm, ....}

\item{selpars}{selection the parameter list containing:
\code{name} is the type of selection to be in the algorithm, ....}

\item{convcrit}{convergence criteria for the algorithm containing:
\code{type} it is the type of convergence criterion; \code{pars} it is the value
the type of convergence criterion; \code{niter} is value of convergence criteria for the algorithm.
\code{stab} is value of stabilization algorithm.}

\item{probpars}{problem the parameter list containing:
\code{name} is problem name to be analyzed;
\code{lim_inf} is lower limit of the information to be generated;
\code{lim_sup} is upper limit of the information to be generated;
\code{opt} other problem parameters.}
}
\value{
result (value) of the optimization.
}
\description{
Is a method that optimizes a problem by iteratively trying to
improve a candidate solution with regard to a given measure of quality
}
\examples{
ExpDE(popsize = 40, mutpars = list(name = "rand", f = 0.2), recpars = list(name = "bin"),
selpars = list(name = "standard"), convcrit = list(types = c("niter", "stab"),
pars = list(niter = 500, nstab = 5)), probpars = list(name = "myfun", lim_inf = -5.12, lim_sup = 5.12, opt = 0))
}
\keyword{de,}
\keyword{optimization}

