% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpDE.R
\name{ExpDE}
\alias{ExpDE}
\title{Experimental Differential Evolution - ExpDE}
\usage{
ExpDE(popsize, mutpars = list(name = "rand", f = 0.2), recpars = list(name =
  "bin"), selpars = list(name = "standard"), stopcrit, probpars)
}
\arguments{
\item{popsize}{population size}

\item{mutpars}{list of named mutation parameters.
See \code{Mutation parameters} for details.}

\item{recpars}{list of named recombination parameters.
See \code{Recombination parameters} for details.}

\item{selpars}{list of named selection parameters.
See \code{Selection parameters} for details.}

\item{stopcrit}{list of named stop criteria parameters. See
\code{Stop criteria} for details.}

\item{probpars}{list of named problem parameters.
   See \code{Problem description} for details.}
}
\value{
A list object containing the final population (sorted by performance)
, the performance vector, and some run statistics.
}
\description{
Modular implementation of the Differential Evolution Algorithm
}
\details{
The detailed description comes here...
"Is a method that optimizes a problem by iteratively trying to
improve a candidate solution with regard to a given measure of
quality via the Differential Evolution algorithm."
}
\section{Mutation parameters}{

}

\section{Recombination parameters}{

}

\section{Selection parameters}{

}

\section{Stop criteria}{

}
\examples{
popsize  <- 40
mutpars  <- list(name = "mutation_rand", f = 0.8)
recpars  <- list(name = "recombination_bin", cr = 0.5, minchange = TRUE)
selpars  <- list(name = "selection_standard")
stopcrit <- list(names = "stop_maxiter", maxiter = 100)
probpars <- list(name   = "rastrigin",
                lim_inf = rep(-5.12,2), lim_sup = rep(5.12,2))

ExpDE(popsize, mutpars, recpars, selpars, stopcrit, probpars)
}
\author{
Felipe Campelo and Moises Botelho
}

