% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exp_params_1.R
\name{get_exp_params_1}
\alias{get_exp_params_1}
\title{Get experimental parameters}
\usage{
get_exp_params_1(ID, echo = TRUE)
}
\arguments{
\item{ID}{a unique positive integer (e.g., a student ID)}

\item{echo}{Boolean, should the function print its output?}
}
\value{
List object with experimental parameters that need to be used for the 
design of the experiment.
}
\description{
Get parameters for practicing the design and running of a comparative experiment 
using ExpDE as a data-generating process.
}
