% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_wgi.R
\name{mutation_wgi}
\alias{mutation_wgi}
\title{/wgi mutation for DE}
\usage{
mutation_wgi(X, mutpars)
}
\arguments{
\item{X}{population matrix}

\item{mutpars}{mutation parameters (see \code{Mutation parameters} for 
details)}
}
\value{
Matrix \code{M} containing the mutated population
}
\description{
Implements the "/wgi/nvecs" mutation (weighted global intermediate) for the 
ExpDE framework. This variant is based on a recombination strategy known as
"weighted global intermediate recombination" (see the References section for 
details)
}
\section{Mutation Parameters}{

The \code{mutpars} parameter contains all parameters required to define the 
mutation. \code{mutation_wgi()} understands the following fields in 
\code{mutpars}:
\itemize{
   \item \code{phi} : scaling factor for difference vector(s).\cr
   Accepts numeric vectors of size 1 or \code{nvecs}.
   \item \code{nvecs} : number of difference vectors to use.\cr 
       Accepts \code{1 <= nvecs <= (nrow(X)/2 - 2)}\cr
       Defaults to 1.
}
}

\section{References}{

Martin Kreutz, Bernhard Sendhoff, Christian Igel, 
"EALib: A C++ class library for evolutionary algorithms". 
Technical Report: http://shark-project.sourceforge.net/2.2.1/doc/EALibManual/EALib.pdf
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br})
}

