% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_arith.R
\name{recombination_arith}
\alias{recombination_arith}
\title{Arithmetic recombination for DE}
\usage{
recombination_arith(X, M, ...)
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{...}{optional parameters (unused)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/arith" (arithmetic) recombination for the ExpDE framework
}
\section{References}{

F. Herrera, M. Lozano, A. M. Sanchez, "A taxonomy for the crossover
operator for real-coded genetic algorithms: an experimental study", 
International Journal of Intelligent Systems 18(3) 309-338, 2003.
}

