% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombination_blxAlpha.R
\name{recombination_blxAlpha}
\alias{recombination_blxAlpha}
\title{Blend Alpha recombination for DE}
\usage{
recombination_blxAlpha(X, M, recpars)
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{recpars}{recombination parameters (see \code{Recombination parameters}
for details)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/blxAlpha" (Blend Alpha) recombination for the ExpDE
framework
}
\section{Recombination Parameters}{

The \code{recpars} parameter contains all parameters required to define the
recombination. \code{recombination_blxAlpha()} understands the following
fields in recpars:
   - \code{alpha} : extrapolation parameter.
   Accepts real value \code{0 <= alpha <= 0.5}. Since the internal coding of
   ExpDE standardizes variables to the interval 0 <= x <= 1, the use of high
   values of \code{alpha} will result in loss of search hability. Lower
   values (e.g., \code{alpha = 0.05}) tend to be more effective.
}

\section{References}{

F. Herrera, M. Lozano, A. M. Sanchez, "A taxonomy for the crossover
operator for real-coded genetic algorithms: an experimental study",
International Journal of Intelligent Systems 18(3) 309-338, 2003.
}

