% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombination_flat.R
\name{recombination_flat}
\alias{recombination_flat}
\title{Flat recombination for DE}
\usage{
recombination_flat(X, M, ...)
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{recpars}{recombination parameters (see \code{Recombination parameters}
for details)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/flat" (Flat) recombination for the ExpDE
framework
}
\section{Recombination Parameters}{

Does not require parameters
}

\section{References}{

Picek, S.; Jakobovic, D.; Golub, M., "On the recombination operator
in the real-coded genetic algorithms," in Evolutionary Computation (CEC),
2013 IEEE Congress on , vol., no., pp.3103-3110, 20-23 June 2013
}

