% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombination_linear.R
\name{recombination_linear}
\alias{recombination_linear}
\title{Linear recombination for DE}
\usage{
recombination_linear(X, M, ...)
}
\arguments{
\item{X}{population matrix (original)}

\item{M}{population matrix (mutated)}

\item{...}{optional parameters (unused)}
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Implements the "/linear" recombination for the ExpDE framework
}
\section{Warning}{

This recombination operator evaluates intermediate candidate solutions,
which adds an extra \code{3*popsize} evaluations per iteration.
}

\section{References}{

F. Herrera, M. Lozano, A. M. Sanchez, "A taxonomy for the crossover
operator for real-coded genetic algorithms: an experimental study",
International Journal of Intelligent Systems 18(3) 309-338, 2003.\cr
A.H. Wright, "Genetic Algorithms for Real Parameter Optimization",
Proc. Foundations of Genetic Algorithms, 205-218, 1991.
}

